+++ src/algorithm/PointLocation.cpp
@@ -20,15 +20,31 @@
 #include <vector>
 
 #include <geos/algorithm/LineIntersector.h>
+#include <geos/algorithm/Orientation.h>
 #include <geos/algorithm/PointLocation.h>
 #include <geos/algorithm/RayCrossingCounter.h>
 #include <geos/geom/CoordinateSequence.h>
 #include <geos/geom/Coordinate.h>
+#include <geos/geom/Envelope.h>
 #include <geos/geom/Location.h>
 #include <geos/util/IllegalArgumentException.h>
 
 namespace geos {
 namespace algorithm { // geos.algorithm
+
+/* public static */
+bool
+PointLocation::isOnSegment(const geom::CoordinateXY& p, const geom::CoordinateXY& p0, const geom::CoordinateXY& p1) 
+{
+    //-- test envelope first since it's faster
+    if (! geom::Envelope::intersects(p0, p1, p))
+        return false;
+    //-- handle zero-length segments
+    if (p.equals2D(p0))
+        return true;
+    bool isOnLine = Orientation::COLLINEAR == Orientation::index(p0, p1, p);
+    return isOnLine;
+}
 
 /* public static */
 bool
@@ -39,13 +55,12 @@ PointLocation::isOnLine(const geom::CoordinateXY& p, const geom::CoordinateSeque
         return false;
     }
 
-    const geom::CoordinateXY* pp = &(pt->getAt<geom::CoordinateXY>(0));
     for(std::size_t i = 1; i < ptsize; ++i) {
-        const geom::CoordinateXY& p1 = pt->getAt<geom::CoordinateXY>(i);
-        if(LineIntersector::hasIntersection(p, *pp, p1)) {
+        if(isOnSegment(p, 
+                        pt->getAt<geom::CoordinateXY>(i - 1), 
+                        pt->getAt<geom::CoordinateXY>(i))) {
             return true;
         }
-        pp = &p1;
     }
     return false;
 }